# Conflicts:
#	FishBun/build/intermediates/incremental/debug/packageDebugResources/compile-file-map.properties
#	FishBun/build/kotlin/compileDebugKotlin/cacheable/last-build.bin
#	FishBun/build/kotlin/compileDebugKotlin/local-state/build-history.bin
#	app/src/main/java/id/worx/device/client/screen/FormBuilder.kt
#	app/src/main/java/id/worx/device/client/screen/main/DetailFormScreen.kt

+++ app/src/main/java/id/worx/device/client/model/SubmitForm.kt
@@ -27,7 +27,7 @@ interface Value{
 
 data class SubmitLocation(
     @SerializedName("address") var address: String? = "",
-    @SerializedName("lat") var lat: Int? = 0,
-    @SerializedName("lng") var lng: Int? = 0
+    @SerializedName("lat") var lat: Double? = 0.0,
+    @SerializedName("lng") var lng: Double? = 0.0
 )
